# perf(ngcc): use `EntryPointManifest` to speed up noop `ProgramBaseEntryPointFinder`