--- a/packages/compiler-cli/ngcc/src/entry_point_finder/entry_point_collector.ts
+++ b/packages/compiler-cli/ngcc/src/entry_point_finder/entry_point_collector.ts
@@ -6,9 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {AbsoluteFsPath, FileSystem, PathSegment} from '../../../src/ngtsc/file_system';
+import {Logger} from '../../../src/ngtsc/logging';
+
 import {EntryPointWithDependencies} from '../dependencies/dependency_host';
 import {DependencyResolver} from '../dependencies/dependency_resolver';
-import {Logger} from '../logging/logger';
 import {NgccConfiguration} from '../packages/configuration';
 import {getEntryPointInfo, IGNORED_ENTRY_POINT, INCOMPATIBLE_ENTRY_POINT, isEntryPoint, NO_ENTRY_POINT} from '../packages/entry_point';
 import {NGCC_DIRECTORY} from '../writing/new_entry_point_file_writer';