… packages/compiler-cli/ngcc/src/entry_point_finder/targeted_entry_point_finder.ts
… packages/compiler-cli/ngcc/src/entry_point_finder/tracing_entry_point_finder.ts
… packages/compiler-cli/ngcc/test/entry_point_finder/directory_walker_entry_point_finder_spec.ts