Previously the `ProgramBasedEntryPointFinder` was parsing all the
entry-points referenced by the program for dependencies even if all the
entry-points had been processed already.

Now this entry-point finder will re-use the `EntryPointManifest` to load
the entry-point dependencies when possible which avoids having to parse
them all again, on every invocation of ngcc.

Previously the `EntryPointManifest` was only used in the
`DirectoryWalkerEntryPointFinder`, which also contained the logic for
computing the contents of the manifest. This logic has been factored out
into an `EntryPointCollector` class. Both the `ProgramBasedEntryPointFinder`
and `DirectoryWalkerEntryPointFinder` now use the `EntryPointManifest` and
the `EntryPointCollector`.

The result of this change is that there is a small cost on the first run of
ngcc to compute and store the manifest - the processing takes 102% of the
processing time before this PR. But on subsequent runs there is a
significant benefit on subsequent runs - the processing takes around 50%
of the processing time before this PR.

---

Using the `super-productivity` as a test environment, we get the following
times for the "hot" (arguably noop) run of `ngcc` using the following command:

```bash
time node node_modules/.bin/ngcc --source node_modules --properties es2015 --first-only --create-ivy-entry-points --use-program-dependencies --async=false --tsconfig src/tsconfig.app.json
```

**Current master:**

(Last row is the average)

User time | System time | CPU usage | Total time
-- | -- | -- | --
2.93 | 0.75 | 106% | 3.465
2.74 | 0.61 | 119% | 2.811
2.8 | 0.62 | 120% | 2.835
2.76 | 0.61 | 119% | 2.818
2.74 | 0.61 | 120% | 2.777
**2.794** | **0.64** | **117%** | **2.9412**

**This PR:**

(Last row is the average)

User time | System time | CPU usage | Total time
-- | -- | -- | --
1.21 | 0.28 | 108% | 1.378
1.24 | 0.28 | 108% | 1.394
1.26 | 0.29 | 110% | 1.398
1.19 | 0.28 | 104% | 1.393
1.17 | 0.28 | 107% | 1.341
1.25 | 0.28 | 101% | 1.511
**1.22** | **0.28** | **106%** | **1.4025**

From these tables it can be seen that the noop (hot) run of ngcc drops from 2.94 to 1.4s (half the time).